WR318527 Make the activity work on its own page without refresh
This entails rewriting almost all the behaviour for workflow
where the forms are being modified in place for editing to
work around Moodle jamming in URL parameters and other things
that weren't an issue when it was on its own page submitting
to its own page after, but inline with AJAX this has issues.

It can't really be broken down smaller since it's all so
interrelated.

Change-Id: I617c1566ce21520eed9d49519a66b0c34543a49c