WR337542: Fix missing buttons on response module
Previsouly only 1 element was found by the css selector, in which case
it was moved to the footer of the card using `.html(content)`.

However, there are now 2 elements which match the selector, one is the
div as before, and another is a paragraph inside the div, so this was
setting the footer content to be the div, then overwritting it again
with the paragraph (which was empty).

This adds in 2 fixes for this, firstly we change to only look for
div elements with the matching id, and secondly, as we are still
in a loop (though there should not only be 1 result) we use append
just in case, so the content doesn't get overwritten if somehow we
find ourselves with multiple matching elements again.

Change-Id: Ic6a65e81d12976336fef1b3e64d7dcfcf658b732