WR302477 Update loading/saving/display
Since most widgets are now supported through the underlying functionality, this
implies the ability to load and save files through submission (e.g. embedding
an image or a recorded WebRTC object), which has to be not only handled during
saving but in all the places this can be displayed through the file system.

Change-Id: I579dc51f3ab1a38da9a4269677712728048f4d80