WR443517 Fix peer results toggle button
The button to switch between "Group" and "All" peer results was not
functioning. The page was always loading showing Group results, and
displaying the "Group" button. This was indicated in the behat tests,
but didn't make sense as clicking the button did nothing when it was
meant to switch to "All" results.

This fixes the behaviour by toggling the button to "All" when Group
results are shown, and vice-versa. It also updates the behat tests to
reflec this behaviour.